update packages, reworking ok for all but memo
