+++ package.json
@@ -9,17 +9,19 @@
   "author": "ERFFFFF",
   "license": "ISC",
   "dependencies": {
-    "@discordjs/opus": "^0.3.2",
-    "discord.js": "^12.2.0",
+    "@discordjs/opus": "^0.3.3",
+    "discord.js": "^12.5.3",
     "express": "^4.17.1",
-    "ffmpeg-static": "^4.2.6",
-    "mongodb": "^3.5.9",
+    "ffmpeg-static": "^4.3.0",
+    "mongodb": "^3.6.9",
     "node-fetch": "^2.6.1",
-    "nodemon": "^2.0.4",
-    "puppeteer": "^5.2.1",
-    "puppeteer-extra": "^3.1.13",
-    "puppeteer-extra-plugin-recaptcha": "^3.1.14",
+    "nodemon": "^2.0.7",
+    "opusscript": "0.0.8",
+    "puppeteer": "^5.5.0",
+    "puppeteer-extra": "^3.1.18",
+    "puppeteer-extra-plugin-recaptcha": "^3.4.0",
     "request": "^2.88.2",
+    "ytdl-core": "^4.8.2",
     "ytdl-core-discord": "^1.3.0"
   },
   "optionalDependencies": {}
